-had to change a bunch of import statements and bring some protocols into their own files since i had linker issues. -share activity VC now displays a visible title. -added limit to total duration a song can have (24 hours). This is enforced during save... -the traditional youtube_id property on the core data song objects is gone. now, the song_id is the actual youtube_id, since the youtube video identifiers are already unique! This helps save space on disk and in memory. Also, the saving of the exact save video into the library more than once is prohibited, and enforced during save. -search bar empty table msg now animated up and down as needed to avoid being blocked by the keyboard. -tab bar can now hide animated...and this feature is being used in a few places. -fixed some minor visual things in the core data error VC. -tried to complete the song editor stuff completely but failed miserably. need to try again (a LOT to fix).

+++ Free Music Library/PlaylistItemTableViewController.m
@@ -145,7 +145,7 @@ static char songIndexPathAssociationKey;  //used to associate cells with images 
                 
                 __weak UIImage *cellImg = albumArt;
                 //calculate how much one length varies from the other.
-                int diff = abs(albumArt.size.width - albumArt.size.height);
+                int diff = abs((int)albumArt.size.width - (int)albumArt.size.height);
                 if(diff > 10){
                     //image is not a perfect (or close to perfect) square. Compensate for this...
                     cellImg = [albumArt imageScaledToFitSize:cell.imageView.frame.size];
